refactor(condo): DOMA-6234 move charts to separate classes & split some chart related interfaces
