--- apps/condo/domains/analytics/components/PaymentChartView.tsx
@@ -1,86 +1,0 @@
-import { Skeleton } from 'antd'
-import ReactECharts from 'echarts-for-react'
-import get from 'lodash/get'
-import React from 'react'
-
-import { useIntl } from '@open-condo/next/intl'
-import { Typography } from '@open-condo/ui'
-import { colors } from '@open-condo/ui/dist/colors'
-
-import { getChartOptions } from '@condo/domains/analytics/utils/helpers'
-
-import PaymentChart from './PaymentChart'
-
-import type { PaymentChartType, PaymentDataType } from './PaymentChart'
-import type { ViewModeTypes } from './TicketChart'
-
-interface IPaymentChartViewProps {
-    data: null | PaymentDataType
-    viewMode: PaymentChartType
-    mapperInstance: PaymentChart
-    loading?: boolean
-    chartConfig: { animationEnabled: boolean, chartOptions?: ReactECharts['props']['opts'] }
-}
-
-const COLOR_SET = [
-    colors.purple['7'],
-    colors.purple['5'],
-    colors.blue['7'],
-    colors.blue['5'],
-    colors.green['7'],
-    colors.green['5'],
-    colors.teal['5'],
-    colors.cyan['5'],
-    colors.cyan['3'],
-]
-
-const EMPTY_CONTAINER_STYLE: React.CSSProperties = {
-    height: '300px',
-    display: 'flex',
-    justifyContent: 'start',
-    alignItems: 'start',
-}
-
-export const PaymentChartView: React.FC<IPaymentChartViewProps> = (props) => {
-    const { viewMode, data, mapperInstance, chartConfig } = props
-
-    const intl = useIntl()
-    const NoData = intl.formatMessage({ id: 'NoData' })
-
-    let legend = [], tooltip = null
-
-    if (data === null) {
-        return <Skeleton loading active paragraph={{ rows: 12 }} />
-    }
-
-    if (data.length === 0) {
-        return (
-            <div style={EMPTY_CONTAINER_STYLE}>
-                <Typography.Text>{NoData}</Typography.Text>
-            </div>
-        )
-    }
-
-    const mapperResult = mapperInstance.getChartConfig(viewMode, data)
-    legend = mapperResult.legend
-    tooltip = mapperResult.tooltip
-
-    const { opts, option } = getChartOptions({
-        legend,
-        tooltip,
-        color: COLOR_SET,
-        viewMode: viewMode as ViewModeTypes,
-        showTitle: false,
-        chartOptions: chartConfig.chartOptions || {},
-        animationEnabled: chartConfig.animationEnabled,
-        series: mapperResult.series,
-        axisData: mapperResult.axisData,
-    })
-    const chartStyle = { height: get(opts, 'height', 'auto') }
-
-    return (
-        <>
-            <ReactECharts opts={opts} option={option} style={chartStyle} notMerge />
-        </>
-    )
-}
