+++ apps/condo/domains/analytics/components/charts/payment/PaymentTotalChart.tsx
@@ -4,7 +4,7 @@ import React, { useMemo } from 'react'
 import { useIntl } from '@open-condo/next/intl'
 import { Typography } from '@open-condo/ui'
 
-import { PaymentChartView } from '@condo/domains/analytics/components/PaymentChartView'
+import { CustomChartView } from '@condo/domains/analytics/components/CustomChartView'
 
 import { PaymentTotalDataMapper } from './dataMappers'
 
@@ -24,7 +24,7 @@ const PaymentTotalChart: IPaymentChartCard = ({ data }) => {
                 <Typography.Title level={3}>{ChartTitle}</Typography.Title>
             </Col>
             <Col span={24}>
-                <PaymentChartView
+                <CustomChartView
                     viewMode='bar'
                     data={data}
                     mapperInstance={dataMapper}
