+++ apps/condo/domains/common/components/BaseChart.ts
@@ -6,10 +6,15 @@ export interface IGetBaseTableConfig<ViewModeType, ResultDataType, RestTableOpti
     (viewMode: ViewModeType, data: ResultDataType, restTableOptions: RestTableOptionsType): BaseTableConfigResult
 }
 
-abstract class BaseChart<ChartConfigType, ChartConfigResult, TableConfigResult, ViewMode, ResultData, RestTableOptions> {
+export abstract class BaseSimpleChart<ChartConfigType, ChartConfigResult, ViewMode, ResultData>{
     constructor (protected chartConfigMap: ChartConfigType) {}
 
     abstract getChartConfig (viewMode: ViewMode, data: ResultData): ChartConfigResult
+}
+
+abstract class BaseChart <ChartConfigType, ChartConfigResult, TableConfigResult, ViewMode, ResultData, RestTableOptions>
+    extends BaseSimpleChart<ChartConfigType, ChartConfigResult, ViewMode, ResultData> {
+
     abstract getTableConfig (viewMode: ViewMode, data: ResultData, restTableOptions: RestTableOptions): TableConfigResult
 }
